+++ GOtravel/GoTravel/Utils/Data/Singleton.swift
