+++ GOtravel/GoTravel/Utils/Data/struct.swift
